Added Reglertest, PI

+++ src/MathLibrary.java
@@ -16,6 +16,14 @@ public class MathLibrary {
     	
     }
     
+    /**
+     * Methode sucht index eines Punktes (Phi) aus einem Double Array (phase).
+     * Gibt ein Integer Array zurück mit 2 Indexwerten, 
+     * index unterhalb (0) und oberhalb (1) des gesuchten Punktes.
+     * @param phase
+     * @param phi
+     * @return [unterer Index, oberer Index]
+     */
     public static int[] int_ver(double[] phase,double phi){
     	int right = phase.length;
     	int left = 0;
@@ -31,26 +39,52 @@ public class MathLibrary {
     	
     }
     
-    
-    public static double[] linspace(int start,int end,int n){
+    /**
+     * Gibt eine linear unterteilte Werteliste als Double Array zurück mit n Punkten
+     * @param start
+     * @param end
+     * @param n
+     * @return lineare Werteliste
+     */
+    public static double[] linspace(double start,double end,int n){
     	double space[] =new double[n];    	
-    	for(double i = 0;i<n;i++) space[(int)i]=start+i*(end-start)/(n-1);
+    	for(int i = 0;i<n;i++) space[i]=start+i*(end-start)/(n-1);
    		return space;   	
     }
     
-    public static double[] linspace(int start,int end){
+	/**
+	 * Gibt eine linear unterteilte Werteliste als Double Array zurück mit 100 Punkten
+	 * @param start
+	 * @param end
+	 * @return lineare Werteliste
+	 */
+    public static double[] linspace(double start,double end){
     	double space[] =new double[100];    	
     	for(double i = 0;i<100;i++) space[(int)i]=start+i*(end-start)/(99);
 		return space;
     	
     }
-    public static double[] logspace(int start,int end,int n){
+    
+    /**
+     * Gibt eine Logarithmisch unterteile Werteliste von 10^start bis und mit 10^ende mit n Anzahl Punkten zurück.
+     * @param start
+     * @param end
+     * @param n
+     * @return Werteliste
+     */
+    public static double[] logspace(double start,double end,int n){
     	double[] space=linspace(start,end,n);
     	for (int i = 0; i < n; i++) space[i]=Math.pow(10, space[i]);
 		return space;    	
     }
     
-    public static double[] logspace(int start,int end){
+    /**
+     * Gibt eine Logarithmisch unterteile Werteliste von 10^start bis und mit10^ende mit 100 Punkten zurück
+     * @param start
+     * @param end
+     * @return Werteliste
+     */
+    public static double[] logspace(double start,double end){
     	double[] space=linspace(start,end,100);
     	for (int i = 0; i < 100; i++) space[i]=Math.pow(10, space[i]);
 		return space;
